Render through Raylib API
